Created functions.py
